+++ test/modules/mod1/mod1.js
@@ -1,0 +1,1 @@
+var div1 ='yes';